refactor: modularize molecule management
