+++ ui/MoleculeCard.js
@@ -1,0 +1,202 @@
+class MoleculeCard {
+    constructor(grid, repository, callbacks = {}) {
+        this.grid = grid;
+        this.repository = repository;
+        this.onDelete = callbacks.onDelete;
+        this.onShowDetails = callbacks.onShowDetails;
+        this.draggedElement = null;
+    }
+
+    createMoleculeCardFromSmiles(smiles, ccdCode) {
+        const card = document.createElement('div');
+        card.className = 'molecule-card';
+        card.draggable = true;
+        card.setAttribute('data-molecule-code', ccdCode);
+
+        const dragHandle = document.createElement('div');
+        dragHandle.className = 'drag-handle';
+        dragHandle.innerHTML = '⋯';
+        card.appendChild(dragHandle);
+
+        const deleteBtn = document.createElement('div');
+        deleteBtn.className = 'delete-btn';
+        deleteBtn.innerHTML = '<svg viewBox="0 0 24 24" width="16" height="16"><path fill="currentColor" d="M19,4H15.5L14.5,3H9.5L8.5,4H5V6H19M6,19A2,2 0 0,0 8,21H16A2,2 0 0,0 18,19V7H6V19Z"/></svg>';
+        deleteBtn.title = `Delete ${ccdCode}`;
+        deleteBtn.addEventListener('click', e => {
+            e.stopPropagation();
+            if (this.onDelete) this.onDelete(ccdCode);
+        });
+        card.appendChild(deleteBtn);
+
+        const codeLabel = document.createElement('div');
+        codeLabel.className = 'molecule-code';
+        codeLabel.textContent = ccdCode;
+        card.appendChild(codeLabel);
+
+        const viewerContainer = document.createElement('div');
+        viewerContainer.className = 'molecule-viewer';
+        viewerContainer.id = `viewer-${ccdCode}`;
+        card.appendChild(viewerContainer);
+
+        const smilesLabel = document.createElement('div');
+        smilesLabel.className = 'smiles-label';
+        smilesLabel.textContent = `SMILES: ${smiles}`;
+        smilesLabel.style.fontSize = '10px';
+        smilesLabel.style.color = '#666';
+        smilesLabel.style.marginTop = '5px';
+        card.appendChild(smilesLabel);
+
+        this.grid.appendChild(card);
+        this.renderSmilesIn2D(smiles, viewerContainer);
+        this.addDragEvents(card);
+    }
+
+    renderSmilesIn2D(smiles, container) {
+        container.innerHTML = `
+            <div style="display:flex;align-items:center;justify-content:center;height:100%;background:#f8f9fa;border:1px solid #e9ecef;border-radius:4px;">
+                <div style="text-align:center;padding:10px;">
+                    <div style="font-size:24px;margin-bottom:5px;">🧪</div>
+                    <div style="font-size:10px;color:#666;">SMILES</div>
+                </div>
+            </div>
+        `;
+    }
+
+    createMoleculeCard(data, ccdCode, format = 'sdf') {
+        const card = document.createElement('div');
+        card.className = 'molecule-card';
+        card.draggable = true;
+        card.setAttribute('data-molecule-code', ccdCode);
+
+        const dragHandle = document.createElement('div');
+        dragHandle.className = 'drag-handle';
+        dragHandle.innerHTML = '⋯';
+        card.appendChild(dragHandle);
+
+        const deleteBtn = document.createElement('div');
+        deleteBtn.className = 'delete-btn';
+        deleteBtn.innerHTML = '<svg viewBox="0 0 24 24" width="16" height="16"><path fill="currentColor" d="M19,4H15.5L14.5,3H9.5L8.5,4H5V6H19M6,19A2,2 0 0,0 8,21H16A2,2 0 0,0 18,19V7H6V19Z"/></svg>';
+        deleteBtn.title = `Delete ${ccdCode}`;
+        deleteBtn.addEventListener('click', e => {
+            e.stopPropagation();
+            if (this.onDelete) this.onDelete(ccdCode);
+        });
+        card.appendChild(deleteBtn);
+
+        const title = document.createElement('h3');
+        title.textContent = ccdCode;
+        title.style.cursor = 'pointer';
+        title.addEventListener('click', () => {
+            if (this.onShowDetails) this.onShowDetails(ccdCode, data, format);
+        });
+        card.appendChild(title);
+
+        const viewerContainer = document.createElement('div');
+        viewerContainer.id = `viewer-${ccdCode}`;
+        viewerContainer.className = 'viewer-container';
+        card.appendChild(viewerContainer);
+
+        this.grid.appendChild(card);
+        this.addDragEvents(card);
+
+        setTimeout(() => {
+            try {
+                const viewer = $3Dmol.createViewer(viewerContainer, { backgroundColor: 'white' });
+                viewer.addModel(data, format);
+                viewer.setStyle({}, { stick: {} });
+                viewer.setStyle({ elem: 'H' }, {});
+                viewer.zoomTo();
+                viewer.render();
+            } catch (e) {
+                console.error(`Error initializing 3Dmol viewer for ${ccdCode}:`, e);
+                viewerContainer.innerHTML = '<p style="text-align:center;padding:20px;">Render Error</p>';
+            }
+        }, 100);
+    }
+
+    createNotFoundCard(ccdCode, message = 'Not found') {
+        const card = document.createElement('div');
+        card.className = 'molecule-card';
+        card.draggable = true;
+        card.setAttribute('data-molecule-code', ccdCode);
+
+        const dragHandle = document.createElement('div');
+        dragHandle.className = 'drag-handle';
+        dragHandle.innerHTML = '⋯';
+        card.appendChild(dragHandle);
+
+        const deleteBtn = document.createElement('div');
+        deleteBtn.className = 'delete-btn';
+        deleteBtn.innerHTML = '<svg viewBox="0 0 24 24" width="16" height="16"><path fill="currentColor" d="M19,4H15.5L14.5,3H9.5L8.5,4H5V6H19M6,19A2,2 0 0,0 8,21H16A2,2 0 0,0 18,19V7H6V19Z"/></svg>';
+        deleteBtn.title = `Delete ${ccdCode}`;
+        deleteBtn.addEventListener('click', e => {
+            e.stopPropagation();
+            if (this.onDelete) this.onDelete(ccdCode);
+        });
+        card.appendChild(deleteBtn);
+
+        const content = document.createElement('div');
+        content.className = 'not-found-content';
+        content.innerHTML = `<h3>${ccdCode}</h3><p>${message}</p>`;
+        card.appendChild(content);
+
+        this.grid.appendChild(card);
+        this.addDragEvents(card);
+    }
+
+    addDragEvents(card) {
+        card.addEventListener('dragstart', e => this.handleDragStart(e));
+        card.addEventListener('dragover', e => this.handleDragOver(e));
+        card.addEventListener('drop', e => this.handleDrop(e));
+        card.addEventListener('dragend', e => this.handleDragEnd(e));
+    }
+
+    handleDragStart(e) {
+        this.draggedElement = e.currentTarget;
+        e.currentTarget.classList.add('dragging');
+        e.dataTransfer.effectAllowed = 'move';
+        e.dataTransfer.setData('text/html', e.currentTarget.outerHTML);
+    }
+
+    handleDragOver(e) {
+        if (e.preventDefault) e.preventDefault();
+        e.dataTransfer.dropEffect = 'move';
+        return false;
+    }
+
+    handleDrop(e) {
+        if (e.stopPropagation) e.stopPropagation();
+        if (this.draggedElement !== e.currentTarget) {
+            const allCards = Array.from(this.grid.querySelectorAll('.molecule-card'));
+            const draggedIndex = allCards.indexOf(this.draggedElement);
+            const targetIndex = allCards.indexOf(e.currentTarget);
+            if (draggedIndex < targetIndex) {
+                e.currentTarget.parentNode.insertBefore(this.draggedElement, e.currentTarget.nextSibling);
+            } else {
+                e.currentTarget.parentNode.insertBefore(this.draggedElement, e.currentTarget);
+            }
+            const molecules = this.repository.molecules;
+            const draggedCode = this.draggedElement.getAttribute('data-molecule-code');
+            const targetCode = e.currentTarget.getAttribute('data-molecule-code');
+            const draggedIdx = molecules.findIndex(m => m.code === draggedCode);
+            const targetIdx = molecules.findIndex(m => m.code === targetCode);
+            if (draggedIdx > -1 && targetIdx > -1) {
+                const [draggedMol] = molecules.splice(draggedIdx, 1);
+                molecules.splice(targetIdx, 0, draggedMol);
+            }
+        }
+        return false;
+    }
+
+    handleDragEnd(e) {
+        e.currentTarget.classList.remove('dragging');
+        this.draggedElement = null;
+    }
+
+    clearAll() {
+        const allCards = this.grid.querySelectorAll('.molecule-card');
+        allCards.forEach(card => card.remove());
+    }
+}
+
+export default MoleculeCard;
